Renamed canonical to device; simplified the Id type
